Reorganized assemblies
